CON-619: Extract highway Fixture for reuse.
